new commit

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,15 +1,15 @@
-#include <stdio.h>
 #include "lists.h"
-
+#include <stdlib.h>
+#include <stdio.h>
+#include <stddef.h>
 /**
-  * print_list - Prints all elements of a list
-  * @h: A linked list
-  *
-  * Return: The number of nodes
-  */
+ * print_list - prints a list
+ * @h: head of list
+ * Return: number of elements
+ */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
@@ -17,10 +17,8 @@ size_t print_list(const list_t *h)
 			printf("[0] (nil)\n");
 		else
 			printf("[%d] %s\n", h->len, h->str);
-
 		h = h->next;
 		count++;
 	}
-
 	return (count);
 }
